update-nukkad-api

+++ nukkad_App/src/main/java/com/code/monks/nukkad/repositories/AddressRepository.java
@@ -8,8 +8,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AddressRepository extends JpaRepository<AddressEntity, Integer> {
-    List<AddressEntity> findAllByCustomerId(int customerId);
-    Optional<AddressEntity> findByIdAndCustomerId(Integer id, int customerId);
+public interface AddressRepository extends JpaRepository<AddressEntity, Long> {
+
+    // Efficiently fetch all addresses for a specific customer
+    List<AddressEntity> findAllByCustomerId(Long customerId);
+
+    // Efficiently fetch address by ID and customerId to ensure ownership
+    Optional<AddressEntity> findByIdAndCustomerId(Long id, Long customerId);
 }
+
